adk-curriculum/step1_researcher/requirements.txt
google-adk==1.26.0
python-dotenv==1.2.2
